backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3
sse-starlette==2.0.0
langdetect
pymupdf==1.24.2
python-docx==1.1.0
